requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
tzdata==2022.1

django-filter~=22.1
fpdf~=1.7.2
reportlab~=3.6.11